web api: use more appropriate data types for preview/pick
before, both used double as a result array type, now the preview uses
float32 (it needs to support a large range of values as it sums them
up), and pick uses the same data type as the original dataset: it does
not need to convert!

+++ src/libertem/web/server.py
@@ -437,7 +437,7 @@ class DataSetPreviewHandler(CORSMixin, tornado.web.RequestHandler):
             )
         log.info("preview futures created")
 
-        full_result = np.zeros(shape=ds.shape[2:])
+        full_result = np.zeros(shape=ds.shape[2:], dtype="float32")
         async for future, result in dd.as_completed(futures, with_results=True):
             for tile in result:
                 tile.copy_to_result(full_result)
@@ -496,7 +496,7 @@ class DataSetPickHandler(CORSMixin, tornado.web.RequestHandler):
                 executor.client.submit(task, **submit_kwargs)
             )
 
-        full_result = np.zeros(shape=ds.shape[2:])
+        full_result = np.zeros(shape=ds.shape[2:], dtype=ds.dtype)
         async for future, result in dd.as_completed(futures, with_results=True):
             for tile in result:
                 tile.copy_to_result(full_result)
